+++ notion_database/cover.py
@@ -1,11 +1,20 @@
+"""
+Notion API Cover
+"""
+from typing import Dict
+
+
 class Cover:
+    """
+    Notion API Cover class
+    """
     def __init__(self):
         """
         init
         """
-        self.result = {}
+        self.result: Dict = {}
 
-    def set_cover_image(self, text=None):
+    def set_cover_image(self, text: str = None):
         """
         Page cover image
         only type of "external" is supported currently