+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## 1.0.0 (2023-5-28)
+
+we've implemented all features, change the version rule to the semantic version.
+
+and changed the license to LGPL 3 in the hope that many people use it.
+
+### Features
+
+* Apply tox, pylint tool
+* Add type-hint code
+* Implement run_query_database function
+* Add support for Python 3.11
+
+### Improvements
+
+* Bump requests from 2.28.2 to 2.31.0
+
+### Deprecate
+
+* Drop support for Python 3.7
+* Deprecate query_database function
+  * move to run_query_database function 
+
 ## 2022-06-28.8 (2023-04-22)
 
 ### Bug Fixes